chmod +x build.sh
little

客户端与服务端断开后，客户端自动重连。

ipc: 新增 sockfd 的 buf 设置；

little

little

ipc 每个域有自己的锁

little

ipc: 客户端的发送放到线程队列中

ipc: 精简代码

little

little

little

little

little

little

消除编译警告

little

little

消除编译警告

little

little

little

little

little

little

little

little

little

little

little

little

little

little

little

little

little

little

ipc：io buf 大小的设置，由 snd 改为 rcv

ipc: 提高 pipe 性能

little

ipc: 新增用于调试的宏定义

threadpool 新增 task_max 功能

little

little

little

little

little

little

little

little

little

little

little

little

新增 log 组件

build log.c

little

little

little

little

little

little

little

little

little

+++ hash/hash.c
@@ -106,7 +106,7 @@ static void hash_free_node(hash_node_t *node)
 static hash_table_t *hash_create_do(size_t height, hash_key_type_t key_type, size_t void_key_len)
 {
 	if (height == 0) {
-		printf("0x2f670a85 args valid(%d:%d:%d) for hash_create!", height, key_type, void_key_len);
+		printf("0x2f670a85 args valid(%ld:%d:%ld) for hash_create!", height, (int)key_type, void_key_len);
 		return NULL;
 	}
 
@@ -153,7 +153,7 @@ void hash_free(hash_table_t *tbl)
 	}
 	/* 数组和表一起释放了，下面无需单独释放 */
 	if (tbl->node_ttl != 0) {
-		printf("0x0cb7eb49 free tbl error ttl: %d", tbl->node_ttl);
+		printf("0x0cb7eb49 free tbl error ttl: %ld", tbl->node_ttl);
 	}
 	free(tbl);
 }
